start sketching some dma stuff
I went around a few circles trying to model the DAC as an object instead of through singleton properties in the packages.

The DMA buffers need to be aligned in memory in order for the ring wrap feature to work. Requesting alignment of individual properties on objects didn't work the first few ways I tried to do it, and seem like it will require knowing more about memory pools or whatever. So for now it's just the singleton version of things.